dojotTester/requirements.txt
gevent==20.12.1
requests==2.24.0
websocket-client==0.57.0
paho-mqtt==1.5.1
pyOpenSSL==21.0.0
